Perform an offline installation of all packages and postpone their configuration
Doing the configuration of the packages while being in system update mode is a rather bad
idea as the postinstall scripts of several services will trigger a restart of the service
which isn't something we want. In system update mode we only have a limited set of
services running and no more.

Signed-off-by: Simon Busch <morphis@gravedo.de>